analysis/cloud functions/close_slots_polling/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery==1.27.2
google-cloud-bigquery-reservation==0.3.0
google-api-core==1.22.1
google-cloud-error-reporting==1.0.0